client: Split out functions to for device
Just part of our long term goal of making client.go smaller
